Remove lots of old code.
