+++ src/test/scala/io/shiftleft/fuzzyc2cpg/passes/CfgCreationPassTests.scala
@@ -2,7 +2,6 @@ package io.shiftleft.fuzzyc2cpg.passes
 
 import better.files.File
 import io.shiftleft.codepropertygraph.Cpg
-import io.shiftleft.fuzzyc2cpg.adapter.{AlwaysEdge, CaseEdge, CfgEdgeType, FalseEdge, TrueEdge}
 import io.shiftleft.passes.IntervalKeyPool
 import org.scalatest.{Matchers, WordSpec}
 import io.shiftleft.semanticcpg.language._
@@ -436,7 +435,7 @@ class CfgFixture(file1Code: String) {
 
   def expected(pairs: (String, CfgEdgeType)*): Set[String] = {
     pairs.map {
-      case (code, cfgEdgeType) =>
+      case (code, _) =>
         codeToNode(code).start.code.head
     }.toSet
   }
